+++ packages/card/src/demos/collapsible.tsx
@@ -36,6 +36,25 @@ export default () => {
       </ProCard>
       <ProCard
         title="Collapsible"
+        headerBordered
+        collapsible="icon"
+        defaultCollapsed
+        onCollapse={(collapse) => console.log(collapse)}
+        extra={
+          <Button
+            size="small"
+            onClick={(e) => {
+              e.stopPropagation();
+            }}
+          >
+            Submit
+          </Button>
+        }
+      >
+        Content
+      </ProCard>
+      <ProCard
+        title="Collapsible"
         bordered
         headerBordered
         collapsible
@@ -97,7 +116,9 @@ export default () => {
           collapsed: buildInCollapsed,
         }: {
           collapsed: boolean;
-        }) => (buildInCollapsed ? <span>Collapse - </span> : <span>Expand - </span>)}
+        }) =>
+          buildInCollapsed ? <span>Collapse - </span> : <span>Expand - </span>
+        }
         style={{ marginBlockStart: 16 }}
         headerBordered
         collapsible