+++ packages/field/src/components/TreeSelect/index.tsx
@@ -140,7 +140,13 @@ const FieldTreeSelect: ProFieldFC<GroupProps> = (
     );
 
     if (render) {
-      return render(rest.text, { mode, ...(fieldProps as any) }, dom) ?? null;
+      return (
+        render(
+          rest.text,
+          { mode, ...(fieldProps as any), treeData: options },
+          dom,
+        ) ?? null
+      );
     }
     return dom;
   }