+++ packages/form/src/components/ModalForm/demos/drawer-form.tsx
@@ -79,7 +79,10 @@ export default () => {
           label="Contract Name"
           placeholder="Please enter a name"
         />
-        <ProFormDateRangePicker name="contractTime" label="Contract Effective Time" />
+        <ProFormDateRangePicker
+          name="contractTime"
+          label="Contract Effective Time"
+        />
       </ProForm.Group>
       <ProForm.Group>
         <ProFormSelect