+++ packages/form/src/components/SchemaForm/index.tsx
@@ -104,7 +104,6 @@ function BetaSchemaForm<T, ValueType = 'text'>(
             'form',
             <LabelIconTip
               label={originItem.title as string}
-              //@ts-expect-error
               tooltip={originItem.tooltip || originItem.tip}
             />,
           );
@@ -126,7 +125,6 @@ function BetaSchemaForm<T, ValueType = 'text'>(
             colProps: originItem.colProps,
             rowProps: originItem.rowProps,
             className: originItem.className,
-            //@ts-expect-error
             tooltip: originItem.tooltip || originItem.tip,
             dependencies: originItem.dependencies,
             proFieldProps: originItem.proFieldProps,
@@ -193,7 +191,7 @@ function BetaSchemaForm<T, ValueType = 'text'>(
     if (columns.length && Array.isArray(columns[0])) return;
     return genItems(columns as ProFormColumnsType<T, ValueType>[]);
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [columns, restProps?.open, action, type, formDomsDeps, !!formRef.current]);
+  }, [columns, restProps?.open, action, type, formDomsDeps]);
 
   /**
    * Append layoutType component specific props