+++ packages/provider/src/locale/zh_HK.tsx
@@ -1,0 +1,74 @@
+export default {
+  moneySymbol: 'HK$',
+  deleteThisLine: '刪除此項',
+  copyThisLine: '複製此項',
+  form: {
+    lightFilter: {
+      more: '更多篩選',
+      clear: '清除',
+      confirm: '確認',
+      itemUnit: '項',
+    },
+  },
+  tableForm: {
+    search: '搜尋', // 「查詢」 → 「搜尋」
+    reset: '重設', // 「重置」 → 「重設」
+    submit: '提交',
+    collapsed: '展開',
+    expand: '收起',
+    inputPlaceholder: '請輸入',
+    selectPlaceholder: '請選擇',
+  },
+  alert: {
+    clear: '取消選取', // 「選擇」 → 「選取」
+    selected: '已選取',
+    item: '項',
+  },
+  pagination: {
+    total: {
+      range: '第',
+      total: '項/總共', // 「條」 → 「項」
+      item: '項',
+    },
+  },
+  tableToolBar: {
+    leftPin: '固定到左邊',
+    rightPin: '固定到右邊',
+    noPin: '不固定',
+    leftFixedTitle: '固定在左側',
+    rightFixedTitle: '固定在右側',
+    noFixedTitle: '不固定',
+    reset: '重設',
+    columnDisplay: '列顯示', // 「列展示」→「列顯示」
+    columnSetting: '列設定', // 「設置」→「設定」
+    fullScreen: '全螢幕', // 「全屏」→「全螢幕」
+    exitFullScreen: '退出全螢幕', // 「退出全屏」→「退出全螢幕」
+    reload: '重新整理', // 「刷新」→「重新整理」
+    density: '密度',
+    densityDefault: '正常',
+    densityLarger: '寬鬆',
+    densityMiddle: '中等',
+    densitySmall: '緊湊',
+  },
+  stepsForm: {
+    next: '下一步',
+    prev: '上一步',
+    submit: '完成',
+  },
+  loginForm: {
+    submitText: '登入',
+  },
+  editableTable: {
+    onlyOneLineEditor: '只能同時編輯一行',
+    action: {
+      save: '保存',
+      cancel: '取消',
+      delete: '刪除',
+      add: '新增一行資料',
+    },
+  },
+  switch: {
+    open: '打開',
+    close: '關閉',
+  },
+};