+++ packages/layout/CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.22.6](https://github.com/ant-design/pro-components/compare/@ant-design/pro-layout@7.22.5...@ant-design/pro-layout@7.22.6) (2025-06-10)
+
+**Note:** Version bump only for package @ant-design/pro-layout
+
+## [7.22.5](https://github.com/ant-design/pro-components/compare/@ant-design/pro-layout@7.22.4...@ant-design/pro-layout@7.22.5) (2025-06-10)
+
+**Note:** Version bump only for package @ant-design/pro-layout
+
+## [7.22.4](https://github.com/ant-design/pro-components/compare/@ant-design/pro-layout@7.22.3...@ant-design/pro-layout@7.22.4) (2025-03-31)
+
+### Bug Fixes
+
+- **tests:** 移除多余的空行以清理代码 ([32cb3b8](https://github.com/ant-design/pro-components/commit/32cb3b8f85d38e5ee60e73440207617340890def))
+
 ## [7.22.3](https://github.com/ant-design/pro-components/compare/@ant-design/pro-layout@7.22.2...@ant-design/pro-layout@7.22.3) (2025-02-17)
 
 ### Bug Fixes